% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{Insert library calls if package is not yet loaded}
\usage{
boot()
}
\value{
library call in script
}
\description{
This function is rarely directly called by the user. More often, it is called
through the \code{>} syntax of the \code{construct_complex} function and
automatically by the \code{\%>>\%} superpipe.
}
