% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean the script of turbokit syntax}
\usage{
clean(detach = TRUE)
}
\arguments{
\item{detach}{Boolean - detach the turbokit package after cleaning?}
}
\value{
Commented turbokit syntax
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Turbokit comes with several features that require special syntax.
Remnants of such syntax are not useful to the data science process. The clean
function is an exit function that transforms any sentences with \emph{pipemath}
format into comments so that it will not be evaluated when running the script.
This is especially helpful for markdown files.

Additionally, it detaches the turbokit package from the namespace.
}
