% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure_turbokit}
\alias{configure_turbokit}
\title{Configure turbokit options through menuing}
\usage{
configure_turbokit()
}
\value{
library call in script
}
\description{
Interactively change package options in an R session. Use \emph{Ctrl + mouse}
to select multiple options.
}
\examples{
# resetting the default
configure_turbokit()
}
