% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{construct_clock}
\alias{construct_clock}
\title{Function to insert clock shortcut combination into R script at cursor position}
\usage{
construct_clock()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert clock shortcut combination into R script at cursor position
}
