% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_cowplot}
\alias{construct_cowplot}
\title{Function to translate cowplot package shortcut combination into R script at
cursor position}
\usage{
construct_cowplot()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate cowplot package shortcut combination into R script at
cursor position
}
