% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_dials}
\alias{construct_dials}
\title{Function to insert dials shortcut combination into R script at cursor position}
\usage{
construct_dials()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to insert dials shortcut combination into R script at cursor position
}
