% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_forcats}
\alias{construct_forcats}
\title{Function to insert forcats shortcut combination into R script at cursor position}
\usage{
construct_forcats()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert forcats shortcut combination into R script at cursor position
}
