% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_ggforce}
\alias{construct_ggforce}
\title{Function to translate ggforce package shortcut combination into R script at
cursor position}
\usage{
construct_ggforce()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate ggforce package shortcut combination into R script at
cursor position
}
