% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_ggmisc}
\alias{construct_ggmisc}
\title{Function to translate several smaller addon packages shortcut combination
into R script at cursor position}
\usage{
construct_ggmisc()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate several smaller addon packages shortcut combination
into R script at cursor position
}
