% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_ggplot}
\alias{construct_ggplot}
\title{Function to insert ggplot shortcut combination into R script at cursor position}
\usage{
construct_ggplot()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to insert ggplot shortcut combination into R script at cursor position
}
