% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_ggraph}
\alias{construct_ggraph}
\title{Function to translate ggraph package shortcut combination into R script at
cursor position}
\usage{
construct_ggraph()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate ggraph package shortcut combination into R script at
cursor position
}
