% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_ggridges}
\alias{construct_ggridges}
\title{Function to translate ggridges package shortcut combination into R script at
cursor position}
\usage{
construct_ggridges()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate ggridges package shortcut combination into R script at
cursor position
}
