% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_ggsci}
\alias{construct_ggsci}
\title{Function to translate ggthemes package shortcut combination into R script at
cursor position}
\usage{
construct_ggsci()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate ggthemes package shortcut combination into R script at
cursor position
}
