% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_parsnip}
\alias{construct_parsnip}
\title{Function to translate parsnip package shortcut combination into R script at
cursor position}
\usage{
construct_parsnip()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to translate parsnip package shortcut combination into R script at
cursor position
}
