% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_readr}
\alias{construct_readr}
\title{Function to insert readr shortcut combination into R script at cursor position}
\usage{
construct_readr()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert readr shortcut combination into R script at cursor position
}
