% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_recipes}
\alias{construct_recipes}
\title{Function to translate recipe package shortcut combination into R script at
cursor position}
\usage{
construct_recipes()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate recipe package shortcut combination into R script at
cursor position
}
