% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_rsample}
\alias{construct_rsample}
\title{Function to insert rsample shortcut combination into R script at cursor position}
\usage{
construct_rsample()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert rsample shortcut combination into R script at cursor position
}
