% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{construct_scales}
\alias{construct_scales}
\title{Function to translate scales package shortcut combination into R script at
cursor position}
\usage{
construct_scales()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate scales package shortcut combination into R script at
cursor position
}
