% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_shiny.R
\name{construct_shiny}
\alias{construct_shiny}
\title{Function to insert shiny shortcut combination into R script at cursor position}
\usage{
construct_shiny()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert shiny shortcut combination into R script at cursor position
}
