% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_stringr}
\alias{construct_stringr}
\title{Function to insert stringr shortcut combination into R script at cursor position}
\usage{
construct_stringr()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert stringr shortcut combination into R script at cursor position
}
