% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_dev.R
\name{construct_testthat}
\alias{construct_testthat}
\title{Function to translate testthat package shortcut combination into R script at
cursor position}
\usage{
construct_testthat()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate testthat package shortcut combination into R script at
cursor position
}
