% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_tune}
\alias{construct_tune}
\title{Function to insert tune shortcut combination into R script at cursor position}
\usage{
construct_tune()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to insert tune shortcut combination into R script at cursor position
}
