% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_dev.R
\name{construct_usethis}
\alias{construct_usethis}
\title{Function to translate usethis package shortcut combination into R script at
cursor position}
\usage{
construct_usethis()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to translate usethis package shortcut combination into R script at
cursor position
}
