% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_workflows}
\alias{construct_workflows}
\title{Function to insert workflows shortcut combination into R script at cursor position}
\usage{
construct_workflows()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert workflows shortcut combination into R script at cursor position
}
