% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{construct_yardstick}
\alias{construct_yardstick}
\title{Function to insert yardstick shortcut combination into R script at cursor position}
\usage{
construct_yardstick()
}
\value{
Adjusted cursor position in R script
}
\description{
Function to insert yardstick shortcut combination into R script at cursor position
}
