% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default}
\alias{default}
\title{Helper to restore default package options used for construct_complex package
abbreviations layout.}
\usage{
default()
}
\value{
library call in script
}
\description{
This function undos the action of the read function.
}
