% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_snippet}
\alias{.add_snippet}
\title{Function to add snippet to a R.snippets file}
\usage{
.add_snippet(name, body)
}
\arguments{
\item{name}{Name of the snippet to be added}

\item{body}{Body of the snippet to be added}
}
\value{
A snippet inserted in the .snippet file
}
\description{
Function to add snippet to a R.snippets file
}
