% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_plot_context}
\alias{.check_plot_context}
\title{Function to detect wrong pipe settings when constructing plot}
\usage{
.check_plot_context(warn = T)
}
\arguments{
\item{warn}{Boolean if warning should be given when changing operators}
}
\value{
Boolean TRUE if plot, geom, coord or facet is on line of R script where cursor is
}
\description{
Function to detect wrong pipe settings when constructing plot
}
