% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.complex_dialog}
\alias{.complex_dialog}
\title{Function that determines input to construct_ functions}
\usage{
.complex_dialog()
}
\value{
character vector of user inputs
}
\description{
Function that determines input to construct_ functions
}
