% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.construct_chain}
\alias{.construct_chain}
\title{Transformation step of dialog box input to be construct_* function-friendly}
\usage{
.construct_chain(input)
}
\arguments{
\item{input}{character string to transform to snippet syntax}
}
\value{
character vector stripped of any numeric indicator
}
\description{
Transformation step of dialog box input to be construct_* function-friendly
}
