% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.construct_snippet}
\alias{.construct_snippet}
\title{Translates tidymath into snippet format.}
\usage{
.construct_snippet(input)
}
\arguments{
\item{input}{a list of tidyverse verbs, each list element is a row in output}
}
\value{
a new entry in r.snippets file called "s"
}
\description{
Translates tidymath into snippet format.
}
