% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.create_expression}
\alias{.create_expression}
\title{Helper to transform character string into turbokit-recognized expression of package expansion}
\usage{
.create_expression(package)
}
\arguments{
\item{package}{package name to transform into expression}
}
\value{
library call in script
}
\description{
Helper to transform character string into turbokit-recognized expression of package expansion
}
