% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_clock_abbreviation}
\alias{.expand_clock_abbreviation}
\title{Function to construct abbreviation into character string}
\usage{
.expand_clock_abbreviation(x)
}
\arguments{
\item{x}{string starting with c}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to construct abbreviation into character string
}
