% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_clock_is}
\alias{.expand_clock_is}
\title{Function to translate abbreviation starting with \code{i} into formatted
character string.}
\usage{
.expand_clock_is(x)
}
\arguments{
\item{x}{character vector, maximum length 5.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{i} into formatted
character string.
}
