% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_clock_set}
\alias{.expand_clock_set}
\title{Function to translate abbreviation starting with \code{s} into formatted
character string.}
\usage{
.expand_clock_set(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{s} into formatted
character string.
}
