% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_default_abbreviation}
\alias{.expand_default_abbreviation}
\title{Function to construct abbreviation into character string}
\usage{
.expand_default_abbreviation(x)
}
\arguments{
\item{x}{character string starting with g.}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to construct abbreviation into character string
}
