% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_forcats_abbreviation}
\alias{.expand_forcats_abbreviation}
\title{Function to construct abbreviation into character string}
\usage{
.expand_forcats_abbreviation(x)
}
\arguments{
\item{x}{string starting with f}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to construct abbreviation into character string
}
