% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_forcats_lvls}
\alias{.expand_forcats_lvls}
\title{Function to translate abbreviation starting with \code{a} into formatted
character string.}
\usage{
.expand_forcats_lvls(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{a} into formatted
character string.
}
