% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{.expand_ggforce_geom}
\alias{.expand_ggforce_geom}
\title{Function to translate ggforce abbreviation starting with a into formatted
character string}
\usage{
.expand_ggforce_geom(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate ggforce abbreviation starting with a into formatted
character string
}
