% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{.expand_ggmisc_position}
\alias{.expand_ggmisc_position}
\title{Function to translate ggmisc abbreviation starting with a into formatted
character string}
\usage{
.expand_ggmisc_position(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate ggmisc abbreviation starting with a into formatted
character string
}
