% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_ggplot_geom}
\alias{.expand_ggplot_geom}
\title{Function to translate abbreviation starting with \code{g} into formatted character string}
\usage{
.expand_ggplot_geom(x)
}
\arguments{
\item{x}{character vector, maximum length 3.}
}
\value{
full length character string of input
}
\description{
Function to translate abbreviation starting with \code{g} into formatted character string
}
