% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_ggplot_scale}
\alias{.expand_ggplot_scale}
\title{Function to translate abbreviation starting with \code{s} into formatted character string.}
\usage{
.expand_ggplot_scale(x)
}
\arguments{
\item{x}{character vector, maximum length 4.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{s} into formatted character string.
}
