% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_ggplot_scale_tertiary}
\alias{.expand_ggplot_scale_tertiary}
\title{Function to identify and transform the third element of abbreviation starting with \code{s}.}
\usage{
.expand_ggplot_scale_tertiary(y)
}
\arguments{
\item{y}{character string of length 1.}
}
\value{
character string of length 1.
}
\description{
Function to identify and transform the third element of abbreviation starting with \code{s}.
}
