% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{.expand_ggsci_abbreviation}
\alias{.expand_ggsci_abbreviation}
\title{Function to transform ggsci abbreviation into character string}
\usage{
.expand_ggsci_abbreviation(x)
}
\arguments{
\item{x}{string starting with r.}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to transform ggsci abbreviation into character string
}
