% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{.expand_ggsci_scale}
\alias{.expand_ggsci_scale}
\title{Function to translate ggsci abbreviation starting with a into formatted
character string}
\usage{
.expand_ggsci_scale(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate ggsci abbreviation starting with a into formatted
character string
}
