% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{.expand_parsnip_abbreviation}
\alias{.expand_parsnip_abbreviation}
\title{Function to transform recipes abbreviation into character string}
\usage{
.expand_parsnip_abbreviation(x)
}
\arguments{
\item{x}{string starting with p}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to transform recipes abbreviation into character string
}
