% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_parsnip_get}
\alias{.expand_parsnip_get}
\title{Function to translate parsnip abbreviation starting with \code{a} into
formatted character string.}
\usage{
.expand_parsnip_get(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate parsnip abbreviation starting with \code{a} into
formatted character string.
}
