% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_readr_abbreviation}
\alias{.expand_readr_abbreviation}
\title{Function to construct abbreviation into character string}
\usage{
.expand_readr_abbreviation(x)
}
\arguments{
\item{x}{string starting with r}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to construct abbreviation into character string
}
