% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidyverse.R
\name{.expand_readr_date}
\alias{.expand_readr_date}
\title{Function to translate abbreviation starting with \code{d} into formatted
character string.}
\usage{
.expand_readr_date(x)
}
\arguments{
\item{x}{character vector, maximum length 3.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{d} into formatted
character string.
}
