% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_readr_format}
\alias{.expand_readr_format}
\title{Function to translate abbreviation starting with \code{f} into formatted
character string.}
\usage{
.expand_readr_format(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{f} into formatted
character string.
}
