% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_recipes_abbreviation}
\alias{.expand_recipes_abbreviation}
\title{Function to transform recipes abbreviation into character string}
\usage{
.expand_recipes_abbreviation(x)
}
\arguments{
\item{x}{string starting with r.}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to transform recipes abbreviation into character string
}
