% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{.expand_recipes_all}
\alias{.expand_recipes_all}
\title{Function to translate recipes abbreviation starting with a into formatted
character string}
\usage{
.expand_recipes_all(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate recipes abbreviation starting with a into formatted
character string
}
