% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_recipes_check}
\alias{.expand_recipes_check}
\title{Function to translate recipes abbreviation starting with c into formatted
character string}
\usage{
.expand_recipes_check(x)
}
\arguments{
\item{x}{character vector, maximum length 3.}
}
\value{
full length character string of input.
}
\description{
Function to translate recipes abbreviation starting with c into formatted
character string
}
