% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_recipes_has}
\alias{.expand_recipes_has}
\title{Function to translate recipes abbreviation starting with h into formatted
character string}
\usage{
.expand_recipes_has(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate recipes abbreviation starting with h into formatted
character string
}
