% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_recipes_update}
\alias{.expand_recipes_update}
\title{Function to translate recipes abbreviation starting with u into formatted
character string}
\usage{
.expand_recipes_update(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate recipes abbreviation starting with u into formatted
character string
}
