% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{.expand_rsample_pretty}
\alias{.expand_rsample_pretty}
\title{Function to translate abbreviation starting with \code{pretty} into formatted
character string.}
\usage{
.expand_rsample_pretty(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate abbreviation starting with \code{pretty} into formatted
character string.
}
