% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{.expand_scales_p}
\alias{.expand_scales_p}
\title{Function to translate scales abbreviation starting with a into formatted
character string}
\usage{
.expand_scales_p(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate scales abbreviation starting with a into formatted
character string
}
