% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_shiny.R
\name{.expand_shiny_default}
\alias{.expand_shiny_default}
\title{Function to construct abbreviation into character string}
\usage{
.expand_shiny_default(x)
}
\arguments{
\item{x}{character string starting with g.}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to construct abbreviation into character string
}
