% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_shiny.R
\name{.expand_shiny_i}
\alias{.expand_shiny_i}
\title{Function to translate abbreviation starting with \code{i} into formatted character string}
\usage{
.expand_shiny_i(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input
}
\description{
Function to translate abbreviation starting with \code{i} into formatted character string
}
