% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_shiny.R
\name{.expand_shiny_r}
\alias{.expand_shiny_r}
\title{Function to translate abbreviation starting with \code{r} into formatted character string}
\usage{
.expand_shiny_r(x)
}
\arguments{
\item{x}{character vector, maximum length 4.}
}
\value{
full length character string of input
}
\description{
Function to translate abbreviation starting with \code{r} into formatted character string
}
