% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_dev.R
\name{.expand_testthat_abbreviation}
\alias{.expand_testthat_abbreviation}
\title{Function to transform usethis abbreviation into character string}
\usage{
.expand_testthat_abbreviation(x)
}
\arguments{
\item{x}{string starting with r.}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to transform usethis abbreviation into character string
}
