% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_dev.R
\name{.expand_testthat_t}
\alias{.expand_testthat_t}
\title{Function to translate testthat abbreviation starting with a into formatted
character string}
\usage{
.expand_testthat_t(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate testthat abbreviation starting with a into formatted
character string
}
