% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_user_defined}
\alias{.expand_user_defined}
\title{Translates and inserts abbreviation of function in user's global environment}
\usage{
.expand_user_defined(x)
}
\arguments{
\item{x}{an abbreviation}
}
\value{
a new entry in r.snippets fall called "s"
}
\description{
Translates and inserts abbreviation of function in user's global environment
}
