% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{.expand_user_pref}
\alias{.expand_user_pref}
\title{Expansion function if user has changed user preferences for construct_complex}
\usage{
.expand_user_pref(x)
}
\arguments{
\item{x}{abbreviation}
}
\value{
Changed turbokit-up options and construct_complex behavior
}
\description{
This function changes the behavior of the construct_complex function by changing
the default package arrangement. Its side effects can be undone by calling
the default function.
}
