% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_dev.R
\name{.expand_usethis_p}
\alias{.expand_usethis_p}
\title{Function to translate usethis abbreviation starting with a into formatted
character string}
\usage{
.expand_usethis_p(x)
}
\arguments{
\item{x}{character vector, maximum length 2.}
}
\value{
full length character string of input.
}
\description{
Function to translate usethis abbreviation starting with a into formatted
character string
}
