% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tidymodels.R
\name{.expand_yardstick_abbreviation}
\alias{.expand_yardstick_abbreviation}
\title{Function to construct abbreviation into character string}
\usage{
.expand_yardstick_abbreviation(x)
}
\arguments{
\item{x}{character string starting with s}
}
\value{
Adjusted cursor position in R script
}
\description{
Function to construct abbreviation into character string
}
