% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_complex_userinput}
\alias{.get_complex_userinput}
\title{Transformation step of dialog box input to be construct_* function-friendly}
\usage{
.get_complex_userinput()
}
\value{
character vector stripped of any numeric indicator
}
\description{
Transformation step of dialog box input to be construct_* function-friendly
}
