% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_snippets_path}
\alias{.get_snippets_path}
\title{Obtain platform-specific snippets file path}
\usage{
.get_snippets_path()
}
\value{
file path
}
\description{
Obtain platform-specific snippets file path
}
