% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.insert}
\alias{.insert}
\title{Function to insert any function into an R script at the cursor position}
\usage{
.insert(verb)
}
\arguments{
\item{verb}{The verb to insert, as a name.}
}
\value{
Verb as function call in R script of the user.
}
\description{
Function to insert any function into an R script at the cursor position
}
