% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.insertmulti}
\alias{.insertmulti}
\title{Function to insert function into an R script at the cursor position}
\usage{
.insertmulti(verb1, verb2)
}
\arguments{
\item{verb1}{The verb to insert, supplied in name form.}

\item{verb2}{A second verb, to be placed within parenthesis of the first verb,
in name form.}
}
\value{
Verb as function call in R script of the user.
}
\description{
Function to insert function into an R script at the cursor position
}
