% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_snippet}
\alias{.prepare_snippet}
\title{Format character input to proper indentation}
\usage{
.prepare_snippet(body)
}
\arguments{
\item{body}{Body of the snippet to be added}
}
\value{
Formatted character of length 1 containing the snippet body
}
\description{
Format character input to proper indentation
}
