% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.read_snippet}
\alias{.read_snippet}
\title{Function to obtain the snippets present in a R.snippets file}
\usage{
.read_snippet(path)
}
\arguments{
\item{path}{filepath to the R.snippets file}
}
\value{
A list of character vectors of length 1 containing snippet bodies and format
}
\description{
Function to obtain the snippets present in a R.snippets file
}
