% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.remove_snippet}
\alias{.remove_snippet}
\title{Function to remove a given snippet from a R.snippets file.}
\usage{
.remove_snippet(name, path)
}
\arguments{
\item{name}{Name of the snippet to be removed}

\item{path}{filepath of the snippet directory}
}
\value{
A renewed R.snippets file without the snippet name specified in the call
}
\description{
Function to remove a given snippet from a R.snippets file.
}
