% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reposition}
\alias{.reposition}
\title{Function to move mouse cursor based on current position}
\usage{
.reposition(n)
}
\arguments{
\item{n}{A number by which to horizontally move the mouse cursor, counting from right to left.}
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to move mouse cursor based on current position
}
