% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reposition_end}
\alias{.reposition_end}
\title{Function to move mouse cursor to the end of the line}
\usage{
.reposition_end()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to move mouse cursor to the end of the line
}
