% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reposition_row}
\alias{.reposition_row}
\title{Function to move mouse cursor to the next row}
\usage{
.reposition_row()
}
\value{
Adjusted cursor position in R script.
}
\description{
Function to move mouse cursor to the next row
}
