% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reposition_str}
\alias{.reposition_str}
\title{Input string at position}
\usage{
.reposition_str(string, inject, index)
}
\arguments{
\item{string}{character string to inject another string in}

\item{inject}{character string to inject into the other string}

\item{index}{position in string where to inject the new string.}
}
\value{
character vector stripped of any numeric indicator.
}
\description{
Input string at position
}
