% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.transform_complex_input}
\alias{.transform_complex_input}
\title{Transformation step of dialog box input to be construct_* function-friendly}
\usage{
.transform_complex_input(x)
}
\arguments{
\item{x}{character string}
}
\value{
character vector stripped of any numeric indicator
}
\description{
Transformation step of dialog box input to be construct_* function-friendly
}
