% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.write_snippet}
\alias{.write_snippet}
\title{Function to write to a R.snippets file}
\usage{
.write_snippet(snippets, path)
}
\arguments{
\item{snippets}{list of character length 1 of snippets to be written to a file.}

\item{path}{file path to the snippet directory}
}
\value{
Renewed R.snippet file.
}
\description{
Function to write to a R.snippets file
}
