% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_visualisation.R
\name{expand_visualisation_default}
\alias{expand_visualisation_default}
\title{Placeholder for default}
\usage{
expand_visualisation_default(x)
}
\arguments{
\item{x}{character string of to compare to default abbreviations}
}
\value{
Adjusted cursor position in R script
}
\description{
Placeholder for default
}
