% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_across}
\alias{insert_across}
\title{across function in an R script with cursor between brackets}
\usage{
insert_across()
}
\value{
across function in an R script with cursor between brackets
}
\description{
across function in an R script with cursor between brackets
}
