% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_case_when}
\alias{insert_case_when}
\title{case_when function in an R script with cursor between brackets}
\usage{
insert_case_when()
}
\value{
case_when function in an R script with cursor between brackets
}
\description{
case_when function in an R script with cursor between brackets
}
