% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_contains}
\alias{insert_contains}
\title{contains function in an R script with cursor between brackets}
\usage{
insert_contains()
}
\value{
contains function in an R script with cursor between brackets
}
\description{
contains function in an R script with cursor between brackets
}
