% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_ends_with}
\alias{insert_ends_with}
\title{ends_with function in an R script with cursor between brackets}
\usage{
insert_ends_with()
}
\value{
ends_with function in an R script with cursor between brackets
}
\description{
ends_with function in an R script with cursor between brackets
}
