% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_everything}
\alias{insert_everything}
\title{everything function in an R script with cursor between brackets}
\usage{
insert_everything()
}
\value{
everything function in an R script with cursor between brackets
}
\description{
everything function in an R script with cursor between brackets
}
