% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_filter}
\alias{insert_filter}
\title{filter function in an R script with cursor between brackets}
\usage{
insert_filter()
}
\value{
filter function in an R script with cursor between brackets
}
\description{
filter function in an R script with cursor between brackets
}
