% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_ggplot}
\alias{insert_ggplot}
\title{ggplot and aes functions in an R script with cursor between bracket of aes}
\usage{
insert_ggplot()
}
\value{
ggplot and aes functions in an R script with cursor between brackets
}
\description{
ggplot and aes functions in an R script with cursor between bracket of aes
}
