% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_group_by}
\alias{insert_group_by}
\title{group_by function in an R script with cursor between brackets}
\usage{
insert_group_by()
}
\value{
group_by function in an R script with cursor between brackets
}
\description{
group_by function in an R script with cursor between brackets
}
