% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_is.numeric}
\alias{insert_is.numeric}
\title{is.numeric function in an R script with cursor between brackets}
\usage{
insert_is.numeric()
}
\value{
is_numeric function in an R script with cursor between brackets
}
\description{
is.numeric function in an R script with cursor between brackets
}
