% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_mutate}
\alias{insert_mutate}
\title{mutate function in an R script with cursor between brackets}
\usage{
insert_mutate()
}
\value{
mutate function in an R script with cursor between brackets
}
\description{
mutate function in an R script with cursor between brackets
}
