% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{insert_pipe}
\alias{insert_pipe}
\title{Insert indented smart pipe operator in R script.}
\usage{
insert_pipe()
}
\value{
\code{\%>\%} or \code{+} in an R script at cursor location,
depending on toggle option, with indenting applied to user settings.
}
\description{
Insert indented smart pipe operator in R script.
}
