% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_pivot_longer}
\alias{insert_pivot_longer}
\title{pivot_longer function in an R script with cursor between brackets}
\usage{
insert_pivot_longer()
}
\value{
pivot_longer function in an R script with cursor between brackets
}
\description{
pivot_longer function in an R script with cursor between brackets
}
