% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_pivot_wider}
\alias{insert_pivot_wider}
\title{pivot_wider function in an R script with cursor between brackets}
\usage{
insert_pivot_wider()
}
\value{
pivot_wider function in an R script with cursor between brackets
}
\description{
pivot_wider function in an R script with cursor between brackets
}
