% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_select}
\alias{insert_select}
\title{select function in an R script with cursor between brackets}
\usage{
insert_select()
}
\value{
select function in an R script with cursor between brackets
}
\description{
select function in an R script with cursor between brackets
}
