% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_starts_with}
\alias{insert_starts_with}
\title{starts_with function in an R script with cursor between brackets}
\usage{
insert_starts_with()
}
\value{
starts_with function in an R script with cursor between brackets
}
\description{
starts_with function in an R script with cursor between brackets
}
