% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_unnest}
\alias{insert_unnest}
\title{nest function in an R script with cursor between brackets}
\usage{
insert_unnest()
}
\value{
nest function in an R script with cursor between brackets
}
\description{
nest function in an R script with cursor between brackets
}
