% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_where}
\alias{insert_where}
\title{where function in an R script with cursor between brackets}
\usage{
insert_where()
}
\value{
where function in an R script with cursor between brackets
}
\description{
where function in an R script with cursor between brackets
}
