% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{pipe_toggle,Turbo-method}
\alias{pipe_toggle,Turbo-method}
\title{Function to get and set pipe operator in options}
\usage{
\S4method{pipe_toggle}{Turbo}(where = NULL, ...)
}
\arguments{
\item{where}{place to insert toggle}

\item{...}{passed down to construct list}
}
\value{
pipe operator options
}
\description{
Function to get and set pipe operator in options
}
