% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\docType{class}
\name{PipeClass-class}
\alias{PipeClass-class}
\alias{PipeClass}
\title{An S4 class for storing toggle options}
\description{
An S4 class for storing toggle options
}
\section{Slots}{

\describe{
\item{\code{options}}{function to get and set options}

\item{\code{value}}{character vector describing operator type}
}}

