% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{toggle}
\alias{toggle}
\title{Get and set pipe operators in package options.}
\usage{
toggle()
}
\value{
Changing pipe operator settings in package options between \code{\%>\%} and \code{+}.
}
\description{
Get and set pipe operators in package options.
}
