% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{toggle_mode}
\alias{toggle_mode}
\title{Get and set turbokit complex input modes}
\usage{
toggle_mode(mode_opt = NULL)
}
\arguments{
\item{mode_opt}{specification of mode by user}
}
\value{
Changing mode settings in package options between \code{tidyverse}, \code{tidymodels}, and \code{shiny}.
}
\description{
Get and set turbokit complex input modes
}
