% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\docType{class}
\name{Turbo-class}
\alias{Turbo-class}
\alias{Turbo}
\title{An S4 class for storing turbokit toggle options}
\description{
An S4 class for storing turbokit toggle options
}
\section{Slots}{

\describe{
\item{\code{pipe_options}}{function to get and set options}

\item{\code{mode_options}}{function to get and set package shortcut mode}

\item{\code{value}}{character vector describing operator type}
}}

