% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turbo}
\alias{turbo}
\title{Full list of turbokit abbreviations}
\format{
A data frame with 917 rows and 4variables:
\describe{
\item{abbreviation}{Abbreviation of a function in a package in a given mode}
\item{fun}{Function corresponding to the abbreviation}
\item{package}{Package where the function lives}
\item{mode}{Turbokit mode which will recognize the abbreviation}
...
}
}
\source{
Donald Seinen
}
\usage{
turbo
}
\description{
A dataset containing a detailed description of all functions currently
recognized by the turbokit package, with their corresponding metadata. Useful
}
\keyword{datasets}
