% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbokit.R
\docType{package}
\name{turbokit}
\alias{turbokit}
\title{turbokit: Speedy tidyverse and pipe input}
\description{
The turbokit package provides two categories of important functions:
insert_ and toggle_.
}
\section{Setting up turbokit}{

Setting up turbokit requires the user to bind shortcuts to function calls.
This can be done by navigating from RStudio:
\strong{Tools -> Addins -> Browse Addins -> Keyboard Shortcuts}

Suggestion for bindings, for minimal hand travel distance and natural combinations.
\tabular{lc}{
Function \tab Recommended Shortcut\cr
toggle \tab \emph{ctrl+shift+/}\cr
pipe \tab \emph{ctrl+shift+.}\cr
select \tab \emph{ctrl+shift+s}\cr
filter \tab \emph{ctrl+shift+f}\cr
mutate \tab \emph{ctrl+shift+m}\cr
summarise \tab \emph{ctrl+shift+z}\cr
across \tab \emph{ctrl+shift+a}\cr
contains \tab \emph{ctrl+shift+f}\cr
starts_with \tab \emph{ctrl+shift+1}\cr
ends_with \tab \emph{ctrl+shift+2}\cr
}
}

\section{insert_ functions}{

The insert_ functions insert a verb or operator in the R script at the current cursor
location, and place the cursor between the function's brackets. The insert_ functions are
designed to be called with shortcuts, allowing for successive shortcut calls.
}

\section{toggle_ functions}{

The toggle_ function provides a shortcut to change package options, which
makes conditional verb insertion possible. This is used to switch between
widely used pipe operators \\%>\\% and + with a button press. This button press
alters the output of another shortcut. This results in a general operator
shortcut button.
}

