% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure_turbokit}
\alias{configure_turbokit}
\title{Configure turbokit options through menuing}
\usage{
configure_turbokit()
}
\value{
library call in script
}
\description{
Interactively toggle between turbokit package options in an R session. When
selected, switches the options to \code{TRUE} or \code{FALSE}, depending on
the initial setting. Use \emph{Ctrl + mouse} to select multiple options at once.
See \code{vignette("turbokit") for explanation of options. }
}
\examples{
# resetting the default
configure_turbokit()
}
