% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{default}
\alias{default}
\title{Restore default package options used for construct_complex package
abbreviations layout.}
\usage{
default()
}
\value{
library call in script
}
\description{
This function undos the effects of the \code{read()} function.
}
\examples{
# resetting the default
default()
}
